Define the CE classes engine-side behind a thin public facade
