+++ src/_lcm/certainty_equivalent.py
@@ -1,36 +1,85 @@
-"""Engine-side certainty-equivalent helpers: power transform pair and resolver.
+"""Certainty-equivalent classes and engine helpers.
 
-The public `lcm.certainty_equivalent` module imports `power_transform`,
-`power_inverse`, and `CE_VALUE_ARG` from here to build its classes; the
-solver seam (`_lcm.regime_building.Q_and_F`) imports
-`resolve_certainty_equivalent` to wire a declared certainty equivalent into
-the backward-induction closure.
+The public `lcm.certainty_equivalent` module re-exports the three classes
+(`CertaintyEquivalent`, `QuasiArithmeticMean`, `PowerMean`) for user code.
+Engine modules may import directly from here.
 """
 
+from abc import ABC, abstractmethod
 from collections.abc import Callable
+from dataclasses import dataclass
 from types import MappingProxyType
-from typing import TYPE_CHECKING
 
+from beartype import beartype
+
+from _lcm.beartype_conf import REGIME_CONF
 from _lcm.utils.functools import get_union_of_args
+from lcm.exceptions import RegimeInitializationError
 from lcm.typing import FloatND
-
-# `lcm.certainty_equivalent` imports the transform pair and `CE_VALUE_ARG` from
-# this module at load time, so this module cannot import the CE classes at module
-# scope without a cycle. `resolve_certainty_equivalent` imports the concrete class
-# lazily inside its body; for the signature, the precise types are kept for ty and
-# a wide `object` runtime form is used for beartype (which cannot resolve the
-# would-be-circular forward references).
-if TYPE_CHECKING:
-    from lcm.certainty_equivalent import CertaintyEquivalent, QuasiArithmeticMean
-
-    _CEInput = CertaintyEquivalent | None
-    _ResolvedCE = QuasiArithmeticMean | None
-else:
-    _CEInput = object
-    _ResolvedCE = object
 
 # Reserved argument name through which transform callables receive values.
 CE_VALUE_ARG = "value"
+
+
+class CertaintyEquivalent(ABC):
+    """Base class for certainty-equivalent specifications.
+
+    Declared on a non-terminal `Regime` via `certainty_equivalent=...`. The
+    engine dispatches on the concrete subclass; `QuasiArithmeticMean` is
+    the shipped implementation. When the field is `None` (the default), the
+    continuation is aggregated as the linear expectation `E[V']`. Only
+    `GridSearch` supports a nonlinear certainty equivalent.
+    """
+
+    @property
+    @abstractmethod
+    def param_names(self) -> frozenset[str]:
+        """Names of the certainty equivalent's runtime parameters."""
+
+
+@beartype(conf=REGIME_CONF)
+@dataclass(frozen=True, kw_only=True)
+class QuasiArithmeticMean(CertaintyEquivalent):
+    """Certainty equivalent `CE = g⁻¹(Σ_r p_r · E_w[g(V'_r)])`.
+
+    A quasi-arithmetic (Kolmogorov) mean: `transform` (`g`) is applied
+    elementwise to next-period values before every expectation - over
+    stochastic state transitions and over regime transitions - and
+    `inverse` (`g⁻¹`) once, after the regime-probability-weighted sum. Both
+    callables take the value array as the reserved first argument `value`;
+    every further signature argument becomes a runtime parameter under the
+    pseudo-function name `certainty_equivalent` in the regime's params
+    (`{"certainty_equivalent": {"<arg>": ...}}`).
+
+    Combined with a user-supplied Bellman aggregator `H` this expresses
+    Epstein-Zin and other transformed-expectation recursive preferences.
+    The parameters are read from the params template only, not from DAG
+    function outputs.
+    """
+
+    transform: Callable[..., FloatND]
+    """`g` — applied elementwise to next-period values before every expectation."""
+
+    inverse: Callable[..., FloatND]
+    """`g⁻¹` — applied once, after the regime-probability-weighted sum."""
+
+    def __post_init__(self) -> None:
+        for name in ("transform", "inverse"):
+            func = getattr(self, name)
+            if CE_VALUE_ARG not in get_union_of_args([func]):
+                msg = (
+                    f"The `{name}` callable of a `QuasiArithmeticMean` must "
+                    f"take the value array via an argument named "
+                    f"'{CE_VALUE_ARG}'."
+                )
+                raise RegimeInitializationError(msg)
+
+    @property
+    def param_names(self) -> frozenset[str]:
+        """Names of the runtime parameters of `transform` and `inverse`."""
+        return frozenset(
+            get_union_of_args([self.transform, self.inverse]) - {CE_VALUE_ARG}
+        )
 
 
 def power_transform(value: FloatND, risk_aversion: FloatND) -> FloatND:
@@ -43,10 +92,26 @@ def power_inverse(value: FloatND, risk_aversion: FloatND) -> FloatND:
     return value ** (1.0 / (1.0 - risk_aversion))
 
 
+@beartype(conf=REGIME_CONF)
+@dataclass(frozen=True, kw_only=True)
+class PowerMean(QuasiArithmeticMean):
+    """Epstein-Zin power-mean certainty equivalent.
+
+    `CE = (E[V'^(1 - risk_aversion)])^(1 / (1 - risk_aversion))` with the
+    runtime parameter `{"certainty_equivalent": {"risk_aversion": ...}}`.
+    Requires strictly positive continuation values; `risk_aversion = 1`
+    (the log case) is not representable. `risk_aversion = 0` reduces to
+    the linear expectation.
+    """
+
+    transform: Callable[..., FloatND] = power_transform
+    inverse: Callable[..., FloatND] = power_inverse
+
+
 def resolve_certainty_equivalent(
-    certainty_equivalent: _CEInput,
+    certainty_equivalent: CertaintyEquivalent | None,
 ) -> tuple[
-    _ResolvedCE,
+    QuasiArithmeticMean | None,
     MappingProxyType[str, str],
     MappingProxyType[str, str],
 ]:
@@ -63,10 +128,6 @@ def resolve_certainty_equivalent(
         arg-to-flat-name mapping.
 
     """
-    # Imported here rather than at module scope to break the load-time cycle with
-    # `lcm.certainty_equivalent` (which imports the transform pair from this module).
-    from lcm.certainty_equivalent import QuasiArithmeticMean  # noqa: PLC0415
-
     if certainty_equivalent is None:
         return None, MappingProxyType({}), MappingProxyType({})
     if not isinstance(certainty_equivalent, QuasiArithmeticMean):
